LayoutTests/platform/glib/imported/w3c/web-platform-tests/websockets/Create-valid-url-protocol-empty.any-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://web-platform.test:49001/echo' failed: WebSocket is closed before the connection is established.

PASS Create WebSocket - wsocket.protocol should be empty before connection is established
